requirements.txt
attrs
pandas
numpy

pyrepresent